requirements.txt
flask
gunicorn
psycopg2-binary
google-generativeai
openai
python-dotenv
flask-cors
jinja2
requests
